Fastighetskartan.qlr
I attach a layer definition file (qlr) for the style of all layers in the Property Map from Lantmäteriet. The style of the layers is made to look as close to that in the lyr-file from Lantmäteriet as possible. In order to use the attached qlr-file, it is just needed to change the source of the layer in plain text according to the users' environment.